src/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.0
pydantic-settings==2.5.2
motor==3.6.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.5
seaborn==0.13.2
scikit-learn==1.5.2
joblib==1.4.2
langchain-cohere==0.3.1
langchain-core==0.3.12
langgraph==0.2.39
